Update random_sample_consensus.py
if using argvs[1] you are only checking the 2nd element of the argvs list, so you will never get a true for argvs[1] == '-sf' for instance (the code checks argvs = '-sf' --> argvs[1] = 's' so argvs[1] == '-sf' is false). By the way, can this library plot in a different color? I have not been able to translate it from the pcl c++ library to the python one.